Removed the MEDS-Extraction components in consideration for pulling those into a separate library.
