--- src/MEDS_transforms/configs/stage_configs/extract_code_metadata.yaml
@@ -1,3 +1,0 @@
-extract_code_metadata:
-  is_metadata: true
-  description_separator: "\n"
