--- src/MEDS_transforms/configs/stage_configs/finalize_MEDS_data.yaml
@@ -1,2 +1,0 @@
-finalize_MEDS_data:
-  do_retype: true
