--- src/MEDS_transforms/configs/stage_configs/finalize_MEDS_metadata.yaml
@@ -1,3 +1,0 @@
-finalize_MEDS_metadata:
-  is_metadata: True
-  do_retype: True
