--- src/MEDS_transforms/configs/stage_configs/merge_to_MEDS_cohort.yaml
@@ -1,3 +1,0 @@
-merge_to_MEDS_cohort:
-  unique_by: "*"
-  additional_sort_by: null
