+++ src/MEDS_transforms/reshard_to_split.py
@@ -3,22 +3,197 @@
 
 import json
 import logging
+import math
 import time
 from collections import defaultdict
+from collections.abc import Sequence
 from functools import partial
 from pathlib import Path
 
 import hydra
+import numpy as np
 import polars as pl
 from meds import subject_id_field, subject_splits_filepath, time_field
 from omegaconf import DictConfig
 
 from MEDS_transforms import PREPROCESS_CONFIG_YAML
-from MEDS_transforms.extract.split_and_shard_subjects import shard_subjects
 from MEDS_transforms.mapreduce.utils import rwlock_wrap, shard_iterator, shuffle_shards
 from MEDS_transforms.utils import stage_init, write_lazyframe
 
 logger = logging.getLogger(__name__)
+
+
+def shard_subjects(
+    subjects: np.ndarray,
+    n_subjects_per_shard: int = 50000,
+    external_splits: dict[str, Sequence[int]] | None = None,
+    split_fracs_dict: dict[str, float] | None = {"train": 0.8, "tuning": 0.1, "held_out": 0.1},
+    seed: int = 1,
+) -> dict[str, list[int]]:
+    """Shard a list of subjects, nested within train/tuning/held-out splits.
+
+    This function takes a list of subjects and shards them into train/tuning/held-out splits, with the shards
+    of a consistent size, nested within the splits. The function will also respect external splits, if
+    provided, such that mandated splits (such as prospective held out sets or pre-existing, task-specific held
+    out sets) are with-held and sharded as separate splits from the IID splits defined by `split_fracs_dict`.
+    It returns a dictionary mapping the split and shard names (realized as f"{split}/{shard}") to the list of
+    subjects in that shard.
+
+    Args:
+        subjects: The list of subjects to shard.
+        n_subjects_per_shard: The maximum number of subjects to include in each shard.
+        external_splits: The externally defined splits to respect. If provided, the keys of this dictionary
+            will be used as split names, and the values as the list of subjects in that split. These
+            pre-defined splits will be excluded from IID splits generated by this function, but will be
+            sharded like normal. Note that this is largely only appropriate for held-out sets for pre-defined
+            tasks or test cases (e.g., prospective tests); training subjects should often still be included in
+            the IID splits to maximize the amount of data that can be used for training.
+        split_fracs_dict: A dictionary mapping the split name to the fraction of subjects to include in that
+            split. Defaults to 80% train, 10% tuning, 10% held-out. This can be None or empty only when
+            external splits fully specify the population.
+        seed: The random seed to use for shuffling the subjects before seeding and sharding. This is useful
+            for ensuring reproducibility.
+
+    Returns:
+        A dictionary mapping f"{split}/{shard}" to the list of subjects in that shard. This may include
+        overlapping subjects across a subset of these splits, but never across shards within a split. Any
+        overlap will solely occur between the an external split and another external split.
+
+    Raises:
+        ValueError: If the sum of the split fractions in `split_fracs_dict` is not equal to 1.
+
+    Examples:
+        >>> subjects = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], dtype=int)
+        >>> shard_subjects(subjects, n_subjects_per_shard=3)
+        {'train/0': [9, 4, 8], 'train/1': [2, 1, 10], 'train/2': [6, 5], 'tuning/0': [3], 'held_out/0': [7]}
+        >>> shard_subjects(subjects, 3, split_fracs_dict={'train': 0.8, 'tuning': 0.2, 'held_out': None})
+        {'train/0': [5, 9, 6], 'train/1': [3, 10, 8], 'train/2': [1, 2], 'tuning/0': [7, 4]}
+        >>> shard_subjects(subjects, 3, split_fracs_dict={'train': 0.8, 'held_out': None})
+        Traceback (most recent call last):
+            ...
+        ValueError: The sum of the split fractions must be equal to 1. Got 0.8 through {'train': 0.8}.
+        >>> external_splits = {
+        ...     'taskA/held_out': np.array([8, 9, 10], dtype=int),
+        ...     'taskB/held_out': np.array([10, 8, 9], dtype=int),
+        ... }
+        >>> shard_subjects(subjects, 3, external_splits)
+        {'train/0': [5, 7, 4],
+         'train/1': [1, 2],
+         'tuning/0': [3],
+         'held_out/0': [6],
+         'taskA/held_out/0': [8, 9, 10],
+         'taskB/held_out/0': [10, 8, 9]}
+        >>> shard_subjects(subjects, n_subjects_per_shard=3, split_fracs_dict={'train': 0.5})
+        Traceback (most recent call last):
+            ...
+        ValueError: The sum of the split fractions must be equal to 1. Got 0.5 through {'train': 0.5}.
+        >>> shard_subjects([1, 2], n_subjects_per_shard=3)
+        Traceback (most recent call last):
+            ...
+        ValueError: Unable to adjust splits to ensure all splits have at least 1 subject.
+        >>> external_splits = {
+        ...     'train': np.array([1, 2, 3, 4, 5, 6], dtype=int),
+        ...     'test': np.array([7, 8, 9, 10], dtype=int),
+        ... }
+        >>> shard_subjects(subjects, 6, external_splits, split_fracs_dict=None)
+        {'train/0': [1, 2, 3, 4, 5, 6], 'test/0': [7, 8, 9, 10]}
+        >>> shard_subjects(subjects, 3, external_splits)
+        {'train/0': [5, 1, 3], 'train/1': [2, 6, 4], 'test/0': [10, 7], 'test/1': [8, 9]}
+    """
+
+    if external_splits is None:
+        external_splits = {}
+    else:
+        for k in list(external_splits.keys()):
+            if not isinstance(external_splits[k], np.ndarray):
+                logger.warning(
+                    f"External split {k} is not a numpy array and thus type safety is not guaranteed. "
+                    f"Attempting to convert to numpy array of dtype {subjects.dtype}."
+                )
+                external_splits[k] = np.array(external_splits[k], dtype=subjects.dtype)
+
+    subjects = np.unique(subjects)
+
+    # Splitting
+    all_external_splits = set().union(*external_splits.values())
+    is_in_external_split = np.isin(subjects, list(all_external_splits))
+    subject_ids_to_split = subjects[~is_in_external_split]
+
+    splits = external_splits
+
+    if split_fracs_dict is not None and None in split_fracs_dict.values():
+        filtered_split_fracs_dict = {k: v for k, v in split_fracs_dict.items() if v is not None}
+        logger.info(
+            "Ignoring splits with null fraction: "
+            + ", ".join(set(split_fracs_dict).difference(filtered_split_fracs_dict))
+        )
+        split_fracs_dict = filtered_split_fracs_dict
+
+    splits_cover = sum(split_fracs_dict.values()) if split_fracs_dict else 0
+
+    rng = np.random.default_rng(seed)
+    if n_subjects := len(subject_ids_to_split):
+        if not math.isclose(splits_cover, 1):
+            raise ValueError(
+                f"The sum of the split fractions must be equal to 1. Got {splits_cover} "
+                f"through {split_fracs_dict}."
+            )
+        split_names_idx = rng.permutation(len(split_fracs_dict))
+        split_names = np.array(list(split_fracs_dict.keys()))[split_names_idx]
+        split_fracs = np.array([split_fracs_dict[k] for k in split_names])
+        split_lens = np.round(split_fracs[:-1] * n_subjects).astype(int)
+        split_lens = np.append(split_lens, n_subjects - split_lens.sum())
+
+        if split_lens.min() == 0:
+            logger.warning(
+                "Some splits are empty. Adjusting splits to ensure all splits have at least 1 subject."
+            )
+            max_split = split_lens.argmax()
+            split_lens[max_split] -= 1
+            split_lens[split_lens.argmin()] += 1
+
+        if split_lens.min() == 0:
+            raise ValueError("Unable to adjust splits to ensure all splits have at least 1 subject.")
+
+        subjects = rng.permutation(subject_ids_to_split)
+        subjects_per_split = np.split(subjects, split_lens.cumsum())
+
+        splits = {**{k: v for k, v in zip(split_names, subjects_per_split)}, **splits}
+    else:
+        if split_fracs_dict:
+            logger.warning(
+                "External splits were provided covering all subjects, but split_fracs_dict was not empty. "
+                "Ignoring the split_fracs_dict."
+            )
+        else:
+            logger.info("External splits were provided covering all subjects.")
+
+    # Sharding
+    final_shards = {}
+    for sp, pts in splits.items():
+        if len(pts) <= n_subjects_per_shard:
+            final_shards[f"{sp}/0"] = pts.tolist()
+        else:
+            pts = rng.permutation(pts)
+            n_pts = len(pts)
+            n_shards = int(np.ceil(n_pts / n_subjects_per_shard))
+            shards = np.array_split(pts, n_shards)
+            for i, shard in enumerate(shards):
+                final_shards[f"{sp}/{i}"] = shard.tolist()
+
+    seen = {}
+
+    for k, pts in final_shards.items():
+        logger.info(f"Split {k} has {len(pts)} subjects.")
+
+        for kk, v in seen.items():
+            shared = set(pts).intersection(v)
+            if shared:
+                logger.info(f"  - intersects {kk} on {len(shared)} subjects.")
+
+        seen[k] = set(pts)
+
+    return final_shards
 
 
 def valid_json_file(fp: Path) -> bool:
