+++ tests/test_add_time_derived_measurements.py
@@ -6,8 +6,8 @@ scripts.
 
 from meds import subject_id_field
 
-from tests.MEDS_Transforms import ADD_TIME_DERIVED_MEASUREMENTS_SCRIPT
-from tests.MEDS_Transforms.transform_tester_base import single_stage_transform_tester
+from tests import ADD_TIME_DERIVED_MEASUREMENTS_SCRIPT
+from tests.transform_tester_base import single_stage_transform_tester
 from tests.utils import parse_meds_csvs
 
 AGE_CALCULATION_STR = """
