+++ tests/test_extract_values.py
@@ -4,8 +4,8 @@ Set the bash env variable `DO_USE_LOCAL_SCRIPTS=1` to use the local py files, ra
 scripts.
 """
 
-from tests.MEDS_Transforms import EXTRACT_VALUES_SCRIPT
-from tests.MEDS_Transforms.transform_tester_base import parse_shards_yaml, single_stage_transform_tester
+from tests import EXTRACT_VALUES_SCRIPT
+from tests.transform_tester_base import parse_shards_yaml, single_stage_transform_tester
 
 INPUT_SHARDS = parse_shards_yaml(
     """
