+++ tests/test_filter_measurements.py
@@ -4,8 +4,8 @@ Set the bash env variable `DO_USE_LOCAL_SCRIPTS=1` to use the local py files, ra
 scripts.
 """
 
-from tests.MEDS_Transforms import FILTER_MEASUREMENTS_SCRIPT
-from tests.MEDS_Transforms.transform_tester_base import single_stage_transform_tester
+from tests import FILTER_MEASUREMENTS_SCRIPT
+from tests.transform_tester_base import single_stage_transform_tester
 from tests.utils import parse_meds_csvs
 
 # This is the code metadata
