+++ tests/test_occlude_outliers.py
@@ -6,8 +6,8 @@ scripts.
 
 import polars as pl
 
-from tests.MEDS_Transforms import OCCLUDE_OUTLIERS_SCRIPT
-from tests.MEDS_Transforms.transform_tester_base import single_stage_transform_tester
+from tests import OCCLUDE_OUTLIERS_SCRIPT
+from tests.transform_tester_base import single_stage_transform_tester
 from tests.utils import MEDS_PL_SCHEMA, parse_meds_csvs
 
 # This is the code metadata
